Revert "feat : Ability to view Audit Log Events for Owned Forms (#6273)"
This reverts commit b38ff695516157a9d4e2acda2f76eef05faeeae2.
