+++ app/(gcforms)/[locale]/(app administration)/admin/(with nav)/events/actions.ts
@@ -1,5 +1,13 @@
 "use server";
-import { logEvent, retrieveEvents } from "@lib/auditLogs";
+import { dynamoDBDocumentClient } from "@lib/integration/awsServicesConnector";
+import {
+  BatchGetCommand,
+  QueryCommand,
+  QueryCommandInput,
+  QueryCommandOutput,
+  BatchGetCommandOutput,
+} from "@aws-sdk/lib-dynamodb";
+import { logEvent } from "@lib/auditLogs";
 import { prisma } from "@lib/integration/prismaConnector";
 import { authorization } from "@lib/privileges";
 import { AccessControlError } from "@lib/auth/errors";
@@ -15,11 +23,104 @@ import {
   pipe,
 } from "valibot";
 
+const _retrieveEvents = async (query: QueryCommandInput) => {
+  const request = new QueryCommand(query);
+
+  const response = (await dynamoDBDocumentClient.send(request)) as QueryCommandOutput;
+  const { Items: eventsIndex, Count: eventsIndexCount } = response;
+
+  if (eventsIndexCount === 0 || eventsIndex === undefined) {
+    return [];
+  }
+  const eventItems = await _retrieveAuditLogs(eventsIndex);
+
+  return eventItems
+    .map((record) => {
+      return {
+        userId: record.UserID,
+        event: record.Event,
+        timestamp: record.TimeStamp,
+        description: record.Description,
+        subject: record.Subject,
+      };
+    })
+    .sort((a, b) => {
+      return b.timestamp - a.timestamp;
+    });
+};
+
+const _retrieveAuditLogs = async (keys: Array<Record<string, string>>) => {
+  let retries = 0;
+  const maxRetries = 3;
+  const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
+  const auditLogs: Array<{
+    UserID: string;
+    Event: string;
+    TimeStamp: number;
+    Description: string;
+    Subject: string;
+  }> = [];
+
+  const batchRequest = new BatchGetCommand({
+    RequestItems: {
+      AuditLogs: {
+        Keys: keys.map((event) => ({
+          UserID: event.UserID,
+          "Event#SubjectID#TimeStamp": event["Event#SubjectID#TimeStamp"],
+        })),
+      },
+    },
+  });
+
+  await dynamoDBDocumentClient.send(batchRequest).then(async (data: BatchGetCommandOutput) => {
+    auditLogs.push(
+      ...(data?.Responses?.AuditLogs?.map((item: Record<string, string | number>) => ({
+        UserID: item.UserID as string,
+        Event: item.Event as string,
+        TimeStamp: item.TimeStamp as number,
+        Description: item.Description as string,
+        Subject: item.Subject as string,
+      })) ?? [])
+    );
+
+    if (data.UnprocessedKeys?.AuditLogs) {
+      while (retries < maxRetries) {
+        // eslint-disable-next-line no-await-in-loop -- Intentional retry logic with delay
+        await delay(200); // Wait for 200ms second before retrying
+        const retryRequest = new BatchGetCommand({
+          RequestItems: {
+            AuditLogs: {
+              Keys: data.UnprocessedKeys.AuditLogs.Keys,
+            },
+          },
+        });
+        const retryResponse: BatchGetCommandOutput =
+          // eslint-disable-next-line no-await-in-loop -- Intentional retry logic
+          await dynamoDBDocumentClient.send(retryRequest);
+        auditLogs.push(
+          ...(retryResponse.Responses?.AuditLogs.map((item: Record<string, string | number>) => ({
+            UserID: item.UserID as string,
+            Event: item.Event as string,
+            TimeStamp: item.TimeStamp as number,
+            Description: item.Description as string,
+            Subject: item.Subject as string,
+          })) ?? [])
+        );
+        if (!retryResponse.UnprocessedKeys?.AuditLogs) {
+          break; // Exit the loop if there are no more unprocessed keys
+        }
+        retries++;
+      }
+    }
+  });
+  return auditLogs;
+};
+
 export const getEventsForUser = async (userId: string) => {
   const {
     user: { id: callingUserId },
   } = await authorization.canViewAllUsers();
-  const events = await retrieveEvents({
+  const events = await _retrieveEvents({
     TableName: "AuditLogs",
     IndexName: "UserByTime",
     Limit: 100,
@@ -43,7 +144,7 @@ export const getEventsForForm = async (formId: string) => {
     user: { id: callingUserId },
   } = await authorization.canViewAllForms();
 
-  const events = await retrieveEvents({
+  const events = await _retrieveEvents({
     TableName: "AuditLogs",
     IndexName: "SubjectByTimestamp",
     Limit: 100,
