+++ i18n/translations/en/form-builder.json
@@ -490,46 +490,6 @@
     "downloadBtnText": "Download form file",
     "title": "Get a copy of your form"
   },
-  "auditDownload": {
-    "description": "Obtain a spreadsheet with the history of all key actions taken in a form, including when a form was created, updated, and archived. Actions can take several minutes to appear in the activity log.",
-    "downloadBtnText": "Download activity log",
-    "title": "Download form activity",
-    "headers": {
-      "formId": "Form ID",
-      "userId": "Team member",
-      "event": "Form action",
-      "timestamp": "Date",
-      "description": "Details"
-    },
-    "events": {
-      "CreateForm": "Form created",
-      "ChangeDeliveryOption": "Delivery option changed",
-      "ChangeSecurityAttribute": "Data classification changed",
-      "ChangeFormPurpose": "Form purpose changed",
-      "ChangeFormSaveAndResume": "Save and resume option changed",
-      "ChangeFormName": "Form title changed",
-      "UpdateForm": "Form updated",
-      "UpdateNotificationsInterval": "Notifications interval updated",
-      "UpdateNotificationsUserSetting": "Notifications user setting updated",
-      "GrantFormAccess": "Granted form access",
-      "RevokeFormAccess": "Revoked form access",
-      "DownloadResponse": "Downloaded responses",
-      "ConfirmResponse": "Confirmed responses",
-      "IdentifyProblemResponse": "Reported problem with responses",
-      "DeleteResponses": "Deleted responses",
-      "RetrieveResponses": "Retrieved responses via API",
-      "CreateAPIKey": "Generated API Key",
-      "DeleteAPIKey": "Deleted API Key",
-      "RefreshAPIKey": "Re-generated API Key",
-      "PublishForm": "Form published"
-    },
-    "errorCreatingDownload": "Error creating download"
-  },
-  "switchToApiMode": {
-    "title": "Switch to API delivery",
-    "description": "New responses will be available through the API endpoint",
-    "btnText": "Switch to API delivery"
-  },
   "formIntroduction": "Form introduction",
   "formInvalidProperty": "Invalid: {{property}}",
   "formMissingProperty": "Missing: {{property}}",
